# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API AFRIGAZ",
        "description": "Documentation API AFRIGAZ",
        "contact": {
            "email": "winnersthec001.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/saleStoreData": {
            "post": {
                "tags": [
                    "Sales"
                ],
                "summary": "Créer une vente avec paiement ou dette",
                "operationId": "e116c7450a1d152a8f99edccedd6cdd1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "branch_id",
                                    "products"
                                ],
                                "properties": {
                                    "branch_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "customer_id": {
                                        "type": "integer",
                                        "example": 1,
                                        "nullable": true
                                    },
                                    "account_id": {
                                        "type": "integer",
                                        "example": 1,
                                        "nullable": true
                                    },
                                    "paid_amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 5000
                                    },
                                    "sale_type": {
                                        "type": "string",
                                        "example": "proforma"
                                    },
                                    "sale_category": {
                                        "type": "string",
                                        "example": "detail"
                                    },
                                    "products": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "product_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "quantity": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "unit_price": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 1000
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Vente créée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Vente enregistrée avec paiement/dette"
                                        },
                                        "sale_id": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "reference": {
                                            "type": "string",
                                            "example": "SALE-20260405120000"
                                        },
                                        "total": {
                                            "type": "number",
                                            "example": 15000
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Stock insuffisant",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Stock insuffisant pour certains produits"
                                        },
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "product_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "message": {
                                                        "type": "string",
                                                        "example": "Stock insuffisant"
                                                    },
                                                    "available": {
                                                        "type": "integer",
                                                        "example": 2
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur interne serveur"
                    }
                }
            }
        },
        "/api/v1/salesGetAllData": {
            "get": {
                "tags": [
                    "Sales"
                ],
                "summary": "Historique complet des ventes",
                "operationId": "4fca0b8bd9b0aa7de227c458d5cc2732",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Nombre de résultats par page",
                        "schema": {
                            "type": "integer",
                            "example": 20
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Numéro de page",
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste paginée des ventes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "reference": {
                                                        "type": "string"
                                                    },
                                                    "transaction_date": {
                                                        "type": "string",
                                                        "format": "date"
                                                    },
                                                    "customer": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer"
                                                            },
                                                            "name": {
                                                                "type": "string"
                                                            }
                                                        },
                                                        "type": "object"
                                                    },
                                                    "user": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer"
                                                            },
                                                            "name": {
                                                                "type": "string"
                                                            }
                                                        },
                                                        "type": "object"
                                                    },
                                                    "items": {
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "id": {
                                                                    "type": "integer"
                                                                },
                                                                "quantity": {
                                                                    "type": "integer"
                                                                },
                                                                "unit_price": {
                                                                    "type": "number"
                                                                },
                                                                "product": {
                                                                    "properties": {
                                                                        "id": {
                                                                            "type": "integer"
                                                                        },
                                                                        "name": {
                                                                            "type": "string"
                                                                        }
                                                                    },
                                                                    "type": "object"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    },
                                                    "total_amount": {
                                                        "type": "number"
                                                    },
                                                    "paid_amount": {
                                                        "type": "number"
                                                    },
                                                    "status": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}